+++ frontend/src/App.css
@@ -18,6 +18,7 @@
   width: 100%;
   margin: 0;
   padding: 0;
+  top: 0;
   color: var(--black);
   background: #fff;
   border-bottom: 2px solid var(--primary-light);
@@ -70,12 +71,7 @@
   box-sizing: border-box;
   z-index: 99;
   border-bottom: 1px solid rgba(32, 33, 36, 0.28);
-  /* box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23); */
   box-shadow: 0 0.75rem 2rem rgba(25, 23, 40, 0.15);
-}
-
-.graph-view {
-  background: #f7f7f7;
 }
 
 .flex-item {
