+++ frontend/src/GraphView/index.js
@@ -1,6 +1,7 @@
 import "./index.css";
 import React, { Component, useEffect, useState } from "react";
 import UserControlPanel from "../UserControlPanel";
+import GraphSummary from "../Graphs/GraphSummary";
 const vetapi = require("../api/vetapi");
 const { formatChemical, intToString } = require("../helpers");
 const {
@@ -54,28 +55,6 @@ class CustomizedXAxisTick extends Component {
     );
   }
 }
-
-// class CustomizedYAxisTick extends Component {
-//   render() {
-//     const { x, y, payload } = this.props;
-//     return (
-//       <g transform={`translate(${x},${y})`}>
-//         <text
-//           fontSize="12"
-//           transform="rotate(35)"
-//           x={0}
-//           y={0}
-//           dy={16}
-//           fill="#FFF"
-//         >
-//           <tspan textAnchor="beginning" x="0" dy="0">
-//             {payload.value}
-//           </tspan>
-//         </text>
-//       </g>
-//     );
-//   }
-// }
 
 function GraphContainer(props) {
   let [graph, setGraph] = useState(null);
@@ -137,7 +116,6 @@ async function TimelineTotal(props) {
             data={res.data}
             margin={{
               top: 5,
-              right: 50,
               left: 50,
               bottom: 5,
             }}
@@ -160,82 +138,6 @@ async function TimelineTotal(props) {
             ></Line>
           </LineChart>
         </ResponsiveContainer>
-      </div>
-    );
-  } catch (err) {
-    handleError(err);
-    return null;
-  }
-}
-
-async function GraphSummary(props) {
-  function format(x) {
-    return x.toLocaleString();
-  }
-
-  try {
-    const { northeast, southwest } = props.viewport;
-    const params = {
-      ne_lat: northeast.lat,
-      ne_lng: northeast.lng,
-      sw_lat: southwest.lat,
-      sw_lng: southwest.lng,
-      year: props.filters.year,
-    };
-    const res = await vetapi.get(`/stats/location/summary`, {
-      params,
-    });
-    const body = (
-      <tbody>
-        <tr>
-          <td>Facilities</td>
-          <td>{res.data["num_facilities"]}</td>
-        </tr>
-        <tr>
-          <td>Distinct Chemicals</td>
-          <td>{res.data["num_distinct_chemicals"]}</td>
-        </tr>
-        <tr>
-          <td>Total Disposal Amount</td>
-          <td>{format(res.data["total"])} lbs</td>
-        </tr>
-        <tr>
-          <td>On-Site Releases</td>
-          <td>{format(res.data["total_on_site"])} lbs</td>
-        </tr>
-        <tr>
-          <td>Off-Site Releases</td>
-          <td>{format(res.data["total_off_site"])} lbs</td>
-        </tr>
-        <tr>
-          <td>Air Releases</td>
-          <td>{format(res.data["total_air"])} lbs</td>
-        </tr>
-        <tr>
-          <td>Water Releases</td>
-          <td>{format(res.data["total_water"])} lbs</td>
-        </tr>
-        <tr>
-          <td>Land Releases</td>
-          <td>{format(res.data["total_land"])} lbs</td>
-        </tr>
-        <tr>
-          <td>Carcinogenic Releases</td>
-          <td>{format(res.data["total_carcinogen"])} lbs</td>
-        </tr>
-      </tbody>
-    );
-    return (
-      <div className="summary">
-        <table>
-          <thead>
-            <tr>
-              <th>Metric</th>
-              <th>Current Location</th>
-            </tr>
-          </thead>
-          {body}
-        </table>
       </div>
     );
   } catch (err) {
@@ -279,7 +181,7 @@ async function GraphTopTenFacilities(props) {
             data={data}
             margin={{
               top: 30,
-              right: 50,
+              // right: 50,
               left: 50,
               bottom: 200,
             }}
@@ -347,7 +249,7 @@ async function GraphTopTenParents(props) {
             // layout="vertical"
             margin={{
               top: 30,
-              right: 50,
+              // right: 50,
               left: 50,
               bottom: 200,
             }}
@@ -410,7 +312,7 @@ async function GraphTopTenChemicals(props) {
             // layout="vertical"
             margin={{
               top: 30,
-              right: 50,
+              // right: 50,
               left: 50,
               bottom: 200,
             }}
@@ -496,7 +398,7 @@ async function TimelineTopFacilities(props) {
             margin={{
               top: 5,
               right: 50,
-              left: 50,
+              // left: 50,
               bottom: 5,
             }}
           >
@@ -591,7 +493,7 @@ async function TimelineTopParents(props) {
             margin={{
               top: 5,
               right: 50,
-              left: 50,
+              // left: 50,
               bottom: 5,
             }}
           >
@@ -684,7 +586,7 @@ async function TimelineTopChemicals(props) {
             margin={{
               top: 5,
               right: 50,
-              left: 50,
+              // left: 50,
               bottom: 5,
             }}
           >
@@ -747,71 +649,79 @@ function GraphView(props) {
 
   return (
     <div className="graph-container" ref={graphRef}>
-      <h1>Location Insights</h1>
-      <div className={`fixed-filter ${visible ? "" : "hidden"}`}>
-        <UserControlPanel
-          chemicals={[]}
-          filters={props.filters}
-          onFilterChange={props.onFilterChange}
-        ></UserControlPanel>
+      <div className="flex-container">
+        <div className="flex-item left">
+          <h1>Location Insights</h1>
+          <div className="filter-container">
+            <UserControlPanel
+              chemicals={[]}
+              filters={props.filters}
+              onFilterChange={props.onFilterChange}
+            ></UserControlPanel>
+          </div>
+          <GraphContainer
+            viewport={props.viewport}
+            filters={props.filters}
+            name="summary"
+            hidden={false}
+            graph={GraphSummary}
+            title="Summary"
+          ></GraphContainer>
+        </div>
+        <div className="flex-item right">
+          <div className="graphs">
+            <GraphContainer
+              viewport={props.viewport}
+              filters={props.filters}
+              name="total_facilities"
+              graph={TimelineTotal}
+              title="Total Releases"
+            ></GraphContainer>
+            <GraphContainer
+              viewport={props.viewport}
+              filters={props.filters}
+              name="total_facilities"
+              graph={GraphTopTenFacilities}
+              title="Total On-Site Releases for Top 10 Facilities (in lbs)"
+            ></GraphContainer>
+            <GraphContainer
+              viewport={props.viewport}
+              filters={props.filters}
+              name="total_parents"
+              graph={GraphTopTenParents}
+              title="Total On-Site Releases for Top 10 Parent Companies (in lbs)"
+            ></GraphContainer>
+            <GraphContainer
+              viewport={props.viewport}
+              filters={props.filters}
+              name="top_graphs"
+              graph={GraphTopTenChemicals}
+              title="Top Ten Chemicals (in lbs)"
+            ></GraphContainer>
+            <GraphContainer
+              viewport={props.viewport}
+              filters={props.filters}
+              name="top_graphs"
+              graph={TimelineTopFacilities}
+              title="Total Releases Over Time for Top Ten Facilities (in lbs)"
+            ></GraphContainer>
+            <GraphContainer
+              viewport={props.viewport}
+              filters={props.filters}
+              name="top_graphs"
+              graph={TimelineTopParents}
+              title="Total Releases Over Time for Top Ten Parent Companies (in lbs)"
+            ></GraphContainer>
+            <GraphContainer
+              viewport={props.viewport}
+              filters={props.filters}
+              name="top_graphs"
+              graph={TimelineTopChemicals}
+              title="Top Ten Chemicals Over Time (in lbs)"
+            ></GraphContainer>
+          </div>
+        </div>
       </div>
-      <GraphContainer
-        viewport={props.viewport}
-        filters={props.filters}
-        name="summary"
-        hidden={false}
-        graph={GraphSummary}
-        title="Summary"
-      ></GraphContainer>
-      <GraphContainer
-        viewport={props.viewport}
-        filters={props.filters}
-        name="total_facilities"
-        graph={TimelineTotal}
-        title="Total Releases"
-      ></GraphContainer>
-      <GraphContainer
-        viewport={props.viewport}
-        filters={props.filters}
-        name="total_facilities"
-        graph={GraphTopTenFacilities}
-        title="Total On-Site Releases for Top 10 Facilities (in lbs)"
-      ></GraphContainer>
-      <GraphContainer
-        viewport={props.viewport}
-        filters={props.filters}
-        name="total_parents"
-        graph={GraphTopTenParents}
-        title="Total On-Site Releases for Top 10 Parent Companies (in lbs)"
-      ></GraphContainer>
-      <GraphContainer
-        viewport={props.viewport}
-        filters={props.filters}
-        name="top_graphs"
-        graph={GraphTopTenChemicals}
-        title="Top Ten Chemicals (in lbs)"
-      ></GraphContainer>
-      <GraphContainer
-        viewport={props.viewport}
-        filters={props.filters}
-        name="top_graphs"
-        graph={TimelineTopFacilities}
-        title="Total Releases Over Time for Top Ten Facilities (in lbs)"
-      ></GraphContainer>
-      <GraphContainer
-        viewport={props.viewport}
-        filters={props.filters}
-        name="top_graphs"
-        graph={TimelineTopParents}
-        title="Total Releases Over Time for Top Ten Parent Companies (in lbs)"
-      ></GraphContainer>
-      <GraphContainer
-        viewport={props.viewport}
-        filters={props.filters}
-        name="top_graphs"
-        graph={TimelineTopChemicals}
-        title="Top Ten Chemicals Over Time (in lbs)"
-      ></GraphContainer>
     </div>
   );
 }
