+++ frontend/src/Graphs/GraphSummary.js
@@ -1,0 +1,80 @@
+import React from "react";
+const vetapi = require("../api/vetapi");
+
+async function GraphSummary(props) {
+  function format(x) {
+    return x.toLocaleString();
+  }
+
+  try {
+    const { northeast, southwest } = props.viewport;
+    const params = {
+      ne_lat: northeast.lat,
+      ne_lng: northeast.lng,
+      sw_lat: southwest.lat,
+      sw_lng: southwest.lng,
+      year: props.filters.year,
+    };
+    const res = await vetapi.get(`/stats/location/summary`, {
+      params,
+    });
+    const body = (
+      <tbody>
+        <tr>
+          <td>Facilities</td>
+          <td>{res.data["num_facilities"]}</td>
+        </tr>
+        <tr>
+          <td>Distinct Chemicals</td>
+          <td>{res.data["num_distinct_chemicals"]}</td>
+        </tr>
+        <tr>
+          <td>Total Disposal Amount</td>
+          <td>{format(res.data["total"])} lbs</td>
+        </tr>
+        <tr>
+          <td>On-Site Releases</td>
+          <td>{format(res.data["total_on_site"])} lbs</td>
+        </tr>
+        <tr>
+          <td>Off-Site Releases</td>
+          <td>{format(res.data["total_off_site"])} lbs</td>
+        </tr>
+        <tr>
+          <td>Air Releases</td>
+          <td>{format(res.data["total_air"])} lbs</td>
+        </tr>
+        <tr>
+          <td>Water Releases</td>
+          <td>{format(res.data["total_water"])} lbs</td>
+        </tr>
+        <tr>
+          <td>Land Releases</td>
+          <td>{format(res.data["total_land"])} lbs</td>
+        </tr>
+        <tr>
+          <td>Carcinogenic Releases</td>
+          <td>{format(res.data["total_carcinogen"])} lbs</td>
+        </tr>
+      </tbody>
+    );
+    return (
+      <div className="summary">
+        <table>
+          <thead>
+            <tr>
+              <th>Metric</th>
+              <th>Current Location</th>
+            </tr>
+          </thead>
+          {body}
+        </table>
+      </div>
+    );
+  } catch (err) {
+    console.log(err);
+    return null;
+  }
+}
+
+export default GraphSummary;
